Revert "Remove redundant temp file when saving"
This reverts commit fc5e231f134932e75ca855b7e587f8e6f81f5953.

+++ src/core/DataFile.cpp
@@ -334,6 +334,7 @@ bool DataFile::writeFile(const QString& filename, bool withResources)
 	const QString fullName = withResources
 		? nameWithExtension(bundleDir + "/" + fInfo.fileName())
 		: nameWithExtension(filename);
+	const QString fullNameTemp = fullName + ".new";
 	const QString fullNameBak = fullName + ".bak";
 
 	using gui::SongEditor;
@@ -379,7 +380,7 @@ bool DataFile::writeFile(const QString& filename, bool withResources)
 		}
 	}
 
-	QSaveFile outfile (fullName);
+	QSaveFile outfile (fullNameTemp);
 
 	if (!outfile.open(QIODevice::WriteOnly | QIODevice::Truncate))
 	{
@@ -418,16 +419,28 @@ bool DataFile::writeFile(const QString& filename, bool withResources)
 		return false;
 	}
 
-
-	// remove old backup file
-	QFile::remove(fullNameBak);
-	if(!ConfigManager::inst()->value( "app", "disablebackup" ).toInt())
+	// make sure the file has been written correctly
+	if( QFileInfo( outfile.fileName() ).size() > 0 )
 	{
-		// move current file to backup file
-		QFile::copy(fullName, fullNameBak);
+		if( ConfigManager::inst()->value( "app", "disablebackup" ).toInt() )
+		{
+			// remove current file
+			QFile::remove( fullName );
+		}
+		else
+		{
+			// remove old backup file
+			QFile::remove( fullNameBak );
+			// move current file to backup file
+			QFile::rename( fullName, fullNameBak );
+		}
+		// move temporary file to current file
+		QFile::rename( fullNameTemp, fullName );
+
+		return true;
 	}
 
-	return true;
+	return false;
 }
